Fe/mails fix (#321)
* fix mails fetching

* refactor mails into its own store
